+++ src/apps/enzyme-finance/ethereum/enzyme-finance.vault.token-fetcher.ts
@@ -1,7 +1,7 @@
 import { Inject } from '@nestjs/common';
 import BigNumber from 'bignumber.js';
 import { gql } from 'graphql-request';
-import _, { compact } from 'lodash';
+import _ from 'lodash';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
@@ -11,6 +11,7 @@ import {
   DataPropsStageParams,
   DisplayPropsStageParams,
   PriceStageParams,
+  UnderlyingTokensStageParams,
 } from '~position/template/app-token.template.position-fetcher';
 import { Network } from '~types/network.interface';
 
@@ -69,47 +70,25 @@ export class EthereumEnzymeFinanceVaultTokenFetcher extends AppTokenTemplatePosi
     return contract.name();
   }
 
-  async getUnderlyingTokenAddresses(contract: EnzymeFinanceVault) {
+  async getUnderlyingTokenAddresses({ contract }: UnderlyingTokensStageParams<EnzymeFinanceVault>) {
     return (await contract.getTrackedAssets()).map(x => x.toLowerCase());
   }
 
-  async getPrice({
-    appToken,
-    contract,
-    tokenLoader,
-  }: PriceStageParams<EnzymeFinanceVault, DefaultDataProps>): Promise<number> {
-    const multicall = this.appToolkit.getMulticall(this.network);
-
-    const decimalsRaw = await this.getDecimals(contract);
-    const supplyRaw = await this.getSupply(contract);
-    const decimals = new BigNumber(10).exponentiatedBy(decimalsRaw);
-    const supply = new BigNumber(supplyRaw.toString()).div(decimals);
-    const underlying = await this.getUnderlyingTokenAddresses(contract);
-
-    const tokenDependencies = await tokenLoader
-      .getMany(underlying.map(tokenAddressRaw => ({ network, address: tokenAddressRaw.toLowerCase() })))
-      .then(deps => compact(deps));
-
+  async getPrice({ appToken, multicall }: PriceStageParams<EnzymeFinanceVault, DefaultDataProps>): Promise<number> {
     const totalAssetUnderManagement = _.sum(
       await Promise.all(
-        underlying.map(async tokenAddressRaw => {
-          const tokenAddress = tokenAddressRaw.toLowerCase();
-          const uTokenContract = this.contractFactory.erc20({ address: tokenAddress, network });
-          const [tokenAmountRaw, decimals] = await Promise.all([
-            multicall.wrap(uTokenContract).balanceOf(appToken.address),
-            multicall.wrap(uTokenContract).decimals(),
-          ]);
-
-          const amount = Number(tokenAmountRaw) / 10 ** decimals;
-          const baseToken = tokenDependencies.find(v => v.address === tokenAddress);
-          if (!baseToken) return 0;
-
-          return baseToken.price * amount;
+        appToken.tokens.map(async token => {
+          const uTokenContract = this.contractFactory.erc20({ address: token.address, network });
+          const tokenAmountRaw = await multicall.wrap(uTokenContract).balanceOf(appToken.address);
+          const amount = Number(tokenAmountRaw) / 10 ** token.decimals;
+          return token.price * amount;
         }),
       ),
     );
 
-    return Number(supply) > 0 ? new BigNumber(totalAssetUnderManagement.toString()).div(supply).toNumber() : 0;
+    return Number(appToken.supply) > 0
+      ? new BigNumber(totalAssetUnderManagement.toString()).div(appToken.supply).toNumber()
+      : 0;
   }
 
   async getDataProps(opts: DataPropsStageParams<EnzymeFinanceVault, DefaultDataProps>): Promise<DefaultDataProps> {