+++ src/apps/hector-network/fantom/hector-network.ws-hec.token-fetcher.ts
@@ -9,6 +9,7 @@ import {
   DisplayPropsStageParams,
   DataPropsStageParams,
   PricePerShareStageParams,
+  UnderlyingTokensStageParams,
 } from '~position/template/app-token.template.position-fetcher';
 import { Network } from '~types/network.interface';
 
@@ -39,11 +40,15 @@ export class FantomHectorNetworkWsHecTokenFetcher extends AppTokenTemplatePositi
     super(appToolkit);
   }
 
+  getContract(address: string): Erc20 {
+    return this.contractFactory.erc20({ address, network: this.network });
+  }
+
   async getAddresses(): Promise<string[]> {
     return ['0x94ccf60f700146bea8ef7832820800e2dfa92eda'];
   }
 
-  async getUnderlyingTokenAddresses(_contract: Erc20) {
+  async getUnderlyingTokenAddresses(_params: UnderlyingTokensStageParams<Erc20>) {
     return '0x75bdef24285013387a47775828bec90b91ca9a5f';
   }